pyproject.toml
[tool.poetry]
name = "get-papers-tool"
version = "0.1.0"
description = ""
authors = ["Mohammad Junaid <mdjunaid34343@gmail.com>"]
readme = "README.md"
packages = [{include = "get_papers_tool", from = "src"}]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
biopython = ">=1.85,<2.0"
typer = ">=0.16.0,<0.17.0"
pandas = ">=2.3.1,<3.0.0"
certifi = "*"

[tool.poetry.scripts]
get-papers-list = "get_papers_tool.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
